prawie all

+++ ZTPProject/Wyniki.cs
@@ -7,7 +7,7 @@ namespace ZTPProject
     public class Wyniki
     {
         public int Id { get; set; }
-        public int result { get; set; }
+        public double result { get; set; }
         public string nick { get; set; }
 
     }
